Moving tests into src/server
