+++ src/server/tests/handlers/cancelHoldHandler.test.ts
@@ -1,16 +1,17 @@
-import handler from "./[id]"
-import { updatePin } from "../helpers"
-import initializePatronTokenAuth from "../../../../src/server/auth"
+import handler from "../../../../pages/api/account/holds/cancel/[id]"
+import { cancelHold } from "../../../../pages/api/account/helpers"
+import initializePatronTokenAuth from "../../auth"
 import type { NextApiRequest, NextApiResponse } from "next"
 
-jest.mock("../../../../src/server/sierraClient")
-jest.mock("../../../../src/server/auth")
-jest.mock("../helpers", () => {
-  const originalModule = jest.requireActual("../helpers")
+jest.mock("../../auth")
+jest.mock("../../../../pages/api/account/helpers", () => {
+  const originalModule = jest.requireActual(
+    "../../../../pages/api/account/helpers"
+  )
   return {
     __esModule: true,
     ...originalModule,
-    updatePin: jest.fn().mockResolvedValueOnce({
+    cancelHold: jest.fn().mockResolvedValueOnce({
       status: "200",
       message: "test",
     }),
@@ -24,7 +25,7 @@ describe("handler", () => {
   beforeEach(() => {
     req = {
       cookies: {},
-      method: "PUT",
+      method: "POST",
       query: {},
       body: {},
     }
@@ -35,35 +36,35 @@ describe("handler", () => {
   })
 
   it("should return 403 if no authenticated patron", async () => {
-    req.query.id = "123456"
+    req.body.patronId = "123456"
     ;(initializePatronTokenAuth as jest.Mock).mockResolvedValueOnce({
       decodedPatron: null,
     })
     await handler(req as NextApiRequest, res as NextApiResponse)
-    expect(updatePin).not.toHaveBeenCalled()
+    expect(cancelHold).not.toHaveBeenCalled()
     expect(res.status).toHaveBeenCalledWith(403)
     expect(res.json).toHaveBeenCalledWith("No authenticated patron")
   })
 
-  it("should return 403 if logged in patron does not match the patron they are updating", async () => {
-    req.query.id = "678910"
+  it("should return 403 if logged in patron does not match the hold they are deleting", async () => {
+    req.body.patronId = "678910"
     ;(initializePatronTokenAuth as jest.Mock).mockResolvedValueOnce({
       decodedPatron: { sub: "123456" },
     })
     await handler(req as NextApiRequest, res as NextApiResponse)
-    expect(updatePin).not.toHaveBeenCalled()
+    expect(cancelHold).not.toHaveBeenCalled()
     expect(res.status).toHaveBeenCalledWith(403)
     expect(res.json).toHaveBeenCalledWith(
-      "Authenticated patron does not match request"
+      "Authenticated patron does not own this hold"
     )
   })
 
-  it("should call updatePin if authentication succeeds", async () => {
-    req.query.id = "123456"
+  it("should call cancelHold if authentication succeeds", async () => {
+    req.body.patronId = "123456"
     ;(initializePatronTokenAuth as jest.Mock).mockResolvedValueOnce({
       decodedPatron: { sub: "123456" },
     })
     await handler(req as NextApiRequest, res as NextApiResponse)
-    expect(updatePin).toHaveBeenCalled()
+    expect(cancelHold).toHaveBeenCalled()
   })
 })
