+++ src/server/tests/handlers/updateHoldHandler.test.ts
@@ -1,18 +1,19 @@
-import handler from "./[id]"
-import { updateSettings } from "../helpers"
-import initializePatronTokenAuth from "../../../../src/server/auth"
+import handler from "../../../../pages/api/account/holds/update/[id]"
+import { updateHold } from "../../../../pages/api/account/helpers"
+import initializePatronTokenAuth from "../../auth"
 import type { NextApiRequest, NextApiResponse } from "next"
 
-jest.mock("../../../../src/server/sierraClient")
-jest.mock("../../../../src/server/auth")
-jest.mock("../helpers", () => {
-  const originalModule = jest.requireActual("../helpers")
+jest.mock("../../auth")
+jest.mock("../../../../pages/api/account/helpers", () => {
+  const originalModule = jest.requireActual(
+    "../../../../pages/api/account/helpers"
+  )
   return {
     __esModule: true,
     ...originalModule,
-    updateSettings: jest.fn().mockResolvedValueOnce({
+    updateHold: jest.fn().mockResolvedValueOnce({
       status: "200",
-      message: "Updated",
+      message: "test",
     }),
   }
 })
@@ -20,11 +21,10 @@ jest.mock("../helpers", () => {
 describe("handler", () => {
   let req: Partial<NextApiRequest>
   let res: Partial<NextApiResponse>
-
   beforeEach(() => {
     req = {
       cookies: {},
-      method: "PUT",
+      method: "POST",
       query: {},
       body: {},
     }
@@ -35,35 +35,36 @@ describe("handler", () => {
   })
 
   it("should return 403 if no authenticated patron", async () => {
-    req.query.id = "123456"
+    req.body.patronId = "123456"
     ;(initializePatronTokenAuth as jest.Mock).mockResolvedValueOnce({
       decodedPatron: null,
     })
     await handler(req as NextApiRequest, res as NextApiResponse)
-    expect(updateSettings).not.toHaveBeenCalled
+    expect(updateHold).not.toHaveBeenCalled()
     expect(res.status).toHaveBeenCalledWith(403)
     expect(res.json).toHaveBeenCalledWith("No authenticated patron")
   })
 
-  it("should return 403 if logged in patron does not match the patron they are updating", async () => {
-    req.query.id = "678910"
+  it("should return 403 if logged in patron does not match the hold they are updating", async () => {
+    req.body.patronId = "678910"
     ;(initializePatronTokenAuth as jest.Mock).mockResolvedValueOnce({
       decodedPatron: { sub: "123456" },
     })
+
     await handler(req as NextApiRequest, res as NextApiResponse)
-    expect(updateSettings).not.toHaveBeenCalled
+    expect(updateHold).not.toHaveBeenCalled()
     expect(res.status).toHaveBeenCalledWith(403)
     expect(res.json).toHaveBeenCalledWith(
-      "Authenticated patron does not match request"
+      "Authenticated patron does not own this hold"
     )
   })
 
-  it("should call updateSettings if authentication succeeds", async () => {
-    req.query.id = "123456"
+  it("should call updateHold if authentication succeeds", async () => {
+    req.body.patronId = "123456"
     ;(initializePatronTokenAuth as jest.Mock).mockResolvedValueOnce({
       decodedPatron: { sub: "123456" },
     })
     await handler(req as NextApiRequest, res as NextApiResponse)
-    expect(updateSettings).toHaveBeenCalled
+    expect(updateHold).toHaveBeenCalled()
   })
 })
